Restructure inheritance for Conversations tests
